requirements.txt
autoscraper==1.1.12
streamlit==0.82.0
pandas==1.0.5
Pillow==8.2.0
